+++ src/main/java/com/inidus/platform/fhir/App.java
@@ -1,4 +1,4 @@
-package com.inidus.platform;
+package com.inidus.platform.fhir;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
