+++ src/main/java/com/inidus/platform/fhir/allergy/AllergyIntoleranceCC.java
@@ -1,4 +1,4 @@
-package com.inidus.platform;
+package com.inidus.platform.fhir.allergy;
 
 import ca.uhn.fhir.model.api.annotation.ResourceDef;
 import org.hl7.fhir.dstu3.model.AllergyIntolerance;
@@ -8,5 +8,5 @@ import org.hl7.fhir.dstu3.model.AllergyIntolerance;
  * Subclass that represents the Care Conect Profile
  */
 @ResourceDef(name="AllergyIntolerance", profile="https://fhir.hl7.org.uk/STU3/StructureDefinition/CareConnect-AllergyIntolerance-1")
-public class CCAllergyIntolerance extends AllergyIntolerance {
+public class AllergyIntoleranceCC extends AllergyIntolerance {
 }
