+++ src/main/java/com/inidus/platform/fhir/openehr/DfText.java
@@ -1,4 +1,4 @@
-package com.inidus.platform.openehr;
+package com.inidus.platform.fhir.openehr;
 
 import org.hl7.fhir.dstu3.model.CodeableConcept;
 import org.hl7.fhir.dstu3.model.Coding;
